requirements.txt
azure-storage-blob==12.2.0
argcomplete==1.11.1
tabulate=0.8.6